requirements.txt
requests
requests-oauthlib
python-twitter-v2
azure-functions